SHIP-0004: Build Environment Variables
This proposal provides a mechanism for developers to append environment
variables to a Build or BuildRun. Many programming languages and
frameworks use environment variables to tune the behavior of their
tooling. For image building tools that aggregate dev frameworks like
Cloud Native Buildpacks, this capaiblity is important because the
number of environment variables that can be tuned is unbounded. Other
developer tools like maven use environment variables to inject
private repository credentials.

This proposal defines the behavior of how environment variables are
defined at the BuildStrategy, Build, and BuildRun levels. It also
proposes enhancements to the shp CLI that make it easier for developers
to set environment variables at build time.