Drop Build Affinity Rules
Affinity and anti-affinity is an incredibly rich and complex API in
Kubernetes. Most of the other scheduler tuning APIs in this SHIP are
simpler to understand and can be wired directly to underlying APIs
(Tekton today, potentially another implementation in the future).
Exposing the entire affinity rule API will make it challenging for
developers to utilize.

We should instead strive for a simpler API that makes it easy for
developers to express their intent. This abstraction can then be
translated to the underlying scheduler APIs. Such an effort is not
trivial and warrants a separate feature/SHIP.

Signed-off-by: Adam Kaplan <adam.kaplan@redhat.com>