Add Project Roadmap
Adding a separate document to hold project roadmap information. This
includes a link to the main GitHub project board we currently use
during backlog refinement sessions.

The roadmap also includes a proposed quarterly release schedule for
2025 that aligns with Tekton's anticipated LTS release schedule (+1
month). The anticipated base Kubernetes and Tekton Pipeline versions
are listed in the `build` sub-project's roadmap. Dates are subject to
change at the discretion of project maintainers.

Signed-off-by: Adam Kaplan <adam.kaplan@redhat.com>