SHIP-0040: Support RuntimeClass in Builds
Proposal to specify the Kubernetes Runtime Class in build pods. This is
a direct follow up to SHIP-0039 that refines our ability to schedule
and execute build pods.

Support for Kubernetes User Namespaces is also floated as an option to
include in the scope, which would fit a broader theme of improving
build security thorugh multiple layers of system isolation. The SHIP is
thus marked as "provisional" to foster discussion.

Signed-off-by: Adam Kaplan <adam.kaplan@redhat.com>